Move distribution tests into assembly tests
