+++ .circleci/config.yml
@@ -164,14 +164,14 @@ jobs:
       - run-bazel-rbe:
           command: bazel test //test-integration/graql/analytics/... --test_output=errors
 
-  test-end-to-end:
+  test-test-assembly:
     machine: true
     working_directory: ~/grakn
     steps:
       - install-bazel-linux-rbe
       - checkout
       - run-bazel-rbe:
-          command: bazel test //test-end-to-end:test-end-to-end --test_output=streamed --spawn_strategy=standalone
+          command: bazel test //test/assembly:test-assembly --test_output=streamed --spawn_strategy=local
 
   test-assembly-mac-zip:
     macos:
@@ -185,7 +185,7 @@ jobs:
       - run: TEST_CACHE_TOKEN=$REPO_GITHUB_TOKEN bazel run @graknlabs_build_tools//ci:test-cache -- --file $(pwd)/bazel-bin/grakn-core-all-mac.zip --get
       - run: unzip bazel-bin/grakn-core-all-mac.zip -d bazel-bin/
       - run: nohup bazel-bin/grakn-core-all-mac/grakn server start
-      - run: bazel test //test/common:grakn-application-test --test_output=streamed --spawn_strategy=standalone --cache_test_results=no
+      - run: bazel test //test/common:grakn-application-test --test_output=streamed --spawn_strategy=local --cache_test_results=no
       - run: bazel-bin/grakn-core-all-mac/grakn server stop
       - run: TEST_CACHE_TOKEN=$REPO_GITHUB_TOKEN bazel run @graknlabs_build_tools//ci:test-cache -- --file $(pwd)/bazel-bin/grakn-core-all-mac.zip --save-success
       - store_artifacts:
@@ -216,7 +216,7 @@ jobs:
       - run: TEST_CACHE_TOKEN=$REPO_GITHUB_TOKEN bazel run @graknlabs_build_tools//ci:test-cache -- --file $(pwd)/bazel-bin/grakn-core-all-linux.tar.gz --get
       - run: tar -xf bazel-bin/grakn-core-all-linux.tar.gz -C bazel-bin
       - run: nohup bazel-bin/grakn-core-all-linux/grakn server start
-      - run: bazel test //test/common:grakn-application-test --test_output=streamed --spawn_strategy=standalone --cache_test_results=no
+      - run: bazel test //test/common:grakn-application-test --test_output=streamed --spawn_strategy=local --cache_test_results=no
       - run: bazel-bin/grakn-core-all-linux/grakn server stop
       - run: TEST_CACHE_TOKEN=$REPO_GITHUB_TOKEN bazel run @graknlabs_build_tools//ci:test-cache -- --file $(pwd)/bazel-bin/grakn-core-all-linux.tar.gz --save-success
       - store_artifacts:
@@ -279,7 +279,7 @@ jobs:
       - run: bazel run //test/deployment/apt:install
       - run: sudo chown -R circleci:circleci /opt/grakn/ # TODO: how do we avoid having to chown?
       - run: nohup grakn server start
-      - run: bazel test //test/common:grakn-application-test --test_output=streamed --spawn_strategy=standalone --cache_test_results=no
+      - run: bazel test //test/common:grakn-application-test --test_output=streamed --spawn_strategy=local --cache_test_results=no
       - run: grakn server stop
       - store_artifacts:
           path: /opt/grakn/core/logs/grakn.log
@@ -486,7 +486,7 @@ workflows:
           filters:
             branches:
               ignore: grakn-release-branch
-      - test-end-to-end:
+      - test-test-assembly:
           filters:
             branches:
               ignore: grakn-release-branch
@@ -505,7 +505,7 @@ workflows:
             - test-integration
             - test-integration-reasoner
             - test-integration-analytics
-            - test-end-to-end
+            - test-test-assembly
       - test-assembly-windows-zip:
           filters:
             branches:
@@ -521,7 +521,7 @@ workflows:
             - test-integration
             - test-integration-reasoner
             - test-integration-analytics
-            - test-end-to-end
+            - test-test-assembly
       - test-assembly-linux-targz:
           filters:
             branches:
@@ -537,7 +537,7 @@ workflows:
             - test-integration
             - test-integration-reasoner
             - test-integration-analytics
-            - test-end-to-end
+            - test-test-assembly
       - test-assembly-docker:
           filters:
             branches:
@@ -553,7 +553,7 @@ workflows:
             - test-integration
             - test-integration-reasoner
             - test-integration-analytics
-            - test-end-to-end
+            - test-test-assembly
       - deploy-apt-snapshot:
           filters:
             branches:
