+++ test/assembly/BUILD
@@ -3,12 +3,12 @@ load("@graknlabs_build_tools//checkstyle:rules.bzl", "checkstyle_test")
 java_test(
     name = "grakn-graql-commands-e2e",
     srcs = [
-        "DistributionE2EConstants.java",
+        "AssemblyE2EConstants.java",
         "GraknGraqlCommandsE2E.java",
     ],
     classpath_resources = ["//test-integration/resources:logback-test"],
     data = ["//:assemble-mac-zip"],
-    test_class = "grakn.core.distribution.GraknGraqlCommandsE2E",
+    test_class = "grakn.core.assembly.GraknGraqlCommandsE2E",
     deps = [
         "//common",
         "//dependencies/maven/artifacts/commons-io",
@@ -21,12 +21,12 @@ java_test(
 java_test(
     name = "grakn-graql-commands-with-a-running-grakn-e2e",
     srcs = [
-        "DistributionE2EConstants.java",
+        "AssemblyE2EConstants.java",
         "GraknGraqlCommands_WithARunningGraknE2E.java",
     ],
     classpath_resources = ["//test-integration/resources:logback-test"],
     data = ["//:assemble-mac-zip"],
-    test_class = "grakn.core.distribution.GraknGraqlCommands_WithARunningGraknE2E",
+    test_class = "grakn.core.assembly.GraknGraqlCommands_WithARunningGraknE2E",
     deps = [
         "//common",
         "//dependencies/maven/artifacts/commons-io",
@@ -34,6 +34,14 @@ java_test(
         "//dependencies/maven/artifacts/org/zeroturnaround:zt-exec",
         "//server",
     ],
+)
+
+test_suite(
+    name = "test-assembly",
+    tests = [
+        "//test/assembly:grakn-graql-commands-e2e",
+        "//test/assembly:grakn-graql-commands-with-a-running-grakn-e2e",
+    ]
 )
 
 checkstyle_test(
